Create process_cost_data.py
